Fixing item arrays on some savegames.
This fixes #1.  Some item lists seem to have an extra 0x01 after
the array count, but before any of the actual array items.  Not
sure where that comes from.  Hopefully adding this in won't
break any existing saves, but it at least works on all the
savegames I've got from when I was writing this.